Initial Commit
This creates all the html, js, and css files. This is done about as is.
There are a couple things that I might want to add like saving
art and stuff like that. I also need to add documentation and
the ability to reset the grid and set the size of the grid.