Makes app look a lot better
I made it look like a real etch a sketch. I also changed the name
to "Scratch-A-Drawing" since I think that Etch-A-Sketch is still
trademarked. This should work for both mobile and desktop,
but the mobile version is not going to be as good since you
can't drag on mobile.

Resolves: #5